# improve version fetching command performance by reading Xcode's version.plist instead of using the `xcodebuild` command