--- a/spec/installed_spec.rb
+++ b/spec/installed_spec.rb
@@ -5,13 +5,13 @@ module XcodeInstall
 
   describe InstalledXcode do
     it 'finds the current Xcode version with whitespace chars' do
-      InstalledXcode.any_instance.expects(:`).with("DEVELOPER_DIR='' \"#{xcode_path}/Contents/Developer/usr/bin/xcodebuild\" -version").returns("Xcode 6.3.1\nBuild version 6D1002")
+      InstalledXcode.any_instance.expects(:`).with("/usr/libexec/PlistBuddy -c \"Print :CFBundleShortVersionString\" \"#{xcode_path}/Contents/version.plist\"").returns('6.3.1')
       installed = InstalledXcode.new(xcode_path)
       installed.version.should == '6.3.1'
     end
 
     it 'is robust against broken Xcode installations' do
-      InstalledXcode.any_instance.expects(:`).with("DEVELOPER_DIR='' \"#{xcode_path}/Contents/Developer/usr/bin/xcodebuild\" -version").returns(nil)
+      InstalledXcode.any_instance.expects(:`).with("/usr/libexec/PlistBuddy -c \"Print :CFBundleShortVersionString\" \"#{xcode_path}/Contents/version.plist\"").returns(nil)
       installed = InstalledXcode.new(xcode_path)
       installed.version.should == '0.0'
     end