closes #273 
Currently It takes a few ms to run `XcodeInstall::Installer.installed_versions`.  The cause seems to be that `xcodebuild` command is executed directly in `fetch_version` method.
Therefore, I refer to #273  and change to parse and read the version.plist inside the Xcode.app to detect Xcode version.

For example, if you run `xcversion list` from CLI and measure the execution time with the `time` command, the execution time changes as follows:

Before improvement:
`xcversion list  3.30s user 2.92s system 40% cpu 15.412 total`
After improvement:
`xcversion list  1.58s user 0.47s system 96% cpu 2.127 total`

As you can see, the execution time has been halved.