@rogerluan 2021-06-15
My results (ran 3-5 times each, pretty consistent results):

```
Before: bundle exec xcversion list  0.93s user 0.48s system 58% cpu 2.407 total
After: bundle exec xcversion list  0.70s user 0.31s system 90% cpu 1.117 total
```

LGTM 🚀